feat: Return old token on duplicate login request
Say the front-end is reloaded. The way it is now, a new request for
login is made to the users controller. And this will add the player
again and generate a new game token without checking the expiration date
of the previous token. The hack for now is the return the old token if
the user is already present in the records. This feature needs to be
tested still.